lightly rework autopatch system; add appropriate annotations to existing patches; standardize
