+++ Source/communityframework/communityframework/DefModExtensions/CompsToAddWhenStuff.cs
